progression-command/progression-command-handler/src/main/java/uk/gov/moj/cpp/progression/command/handler/CompleteSendingSheetHandler.java
package uk.gov.moj.cpp.progression.command.handler;

import uk.gov.justice.services.core.annotation.Component;
import uk.gov.justice.services.core.annotation.Handles;
import uk.gov.justice.services.core.annotation.ServiceComponent;
import uk.gov.justice.services.eventsourcing.source.core.exception.EventStreamException;
import uk.gov.justice.services.messaging.JsonEnvelope;

@ServiceComponent(Component.COMMAND_HANDLER)
public class CompleteSendingSheetHandler extends CaseProgressionCommandHandler {

    @Handles("progression.command.complete-sending-sheet")
    public void completeSendingSheet(final JsonEnvelope command) throws EventStreamException {
        applyToCaseProgressionAggregate(command,
                aCase -> aCase.completeSendingSheet(command));
    }
}
